TestTakerDeliveryLog deleted
