+++ README.md
@@ -1,23 +1,2 @@
 # extension-tao-monitoring
 ------
-
-Collections of the **TAO** statistics
- 
-### TestTaker delivery log
-#### Aggregate statistic for deliveries of the test taker
-
-**Format**:
-
-| test_taker |  nb_executions | nb_finished |
-| ---------- |:-------------:|:-----------:|
-| 0271a3b    | 2             | 2           |
-| 0811a13b   | 2             | 0           |
-| 0121a25b   | 1             | 1           |
-
-*Where*:
-
- - **test_taker** - Login of the test taker
- - **nb_executions** - Count of the started executions
- - **nb_finished** - Count of the finished executions
- 
-> In RDS will be created table with statistics
