--- config/header/TestTakerDeliveryLog.conf.php
@@ -1,6 +1,0 @@
-<?php
-/**
- * Configuration for logger service of the test taker deliveries 
- *
- * Service is used to store data in an efficient way
- */
