+++ scripts/update/Updater.php
@@ -28,8 +28,7 @@ use oat\taoMonitoring\model\DeliveryLog\DeliveryLogService;
 use oat\taoMonitoring\model\TestTakerDeliveryActivityLog\TestTakerDeliveryActivityLogService;
 use oat\taoMonitoring\scripts\install\RegisterRdsDeliveryLog;
 use oat\taoMonitoring\scripts\install\RegisterRdsTestTakerDeliveryActivityLog;
-use oat\taoMonitoring\scripts\update\v0_1_0\DropNBItemColumnFromTtDeliveryLog;
-
+use oat\taoMonitoring\scripts\update\v0_1_0\DropTestTakerDeliveryLogTable;
 
 class Updater extends common_ext_ExtensionUpdater {
 
@@ -57,8 +56,6 @@ class Updater extends common_ext_ExtensionUpdater {
             $this->setVersion('0.0.2');
         }
 
-        $this->skip('0.0.2', '0.1.0');
-
         if ($this->isVersion('0.0.2')) {
 
             // delete NB_ITEM from the monitoring of the test takers deliveries
@@ -69,10 +66,23 @@ class Updater extends common_ext_ExtensionUpdater {
                 'oat\\taoQtiTest\\models\\event\\QtiMoveEvent',
                 array('\\oat\\taoMonitoring\\model\\TestTakerDeliveryLog\\event\\Events', 'qtiMoveEvent')
             );
+
+            // count executions
+            $eventManager->detach(
+                'oat\\taoDelivery\\models\\classes\\execution\\event\\DeliveryExecutionCreated',
+                array('\\oat\\taoMonitoring\\model\\TestTakerDeliveryLog\\event\\Events', 'deliveryExecutionCreated')
+            );
+
+            // finished executions
+            $eventManager->detach(
+                'oat\\taoDelivery\\models\\classes\\execution\\event\\DeliveryExecutionState',
+                array('\\oat\\taoMonitoring\\model\\TestTakerDeliveryLog\\event\\Events', 'deliveryExecutionState')
+            );
+
             $this->getServiceManager()->register(EventManager::CONFIG_ID, $eventManager);
 
             // delete from storage
-            $action = new DropNBItemColumnFromTtDeliveryLog();
+            $action = new DropTestTakerDeliveryLogTable();
             $action([]);
 
             $this->setVersion('0.1.0');
